Adicionando o beginTransaction nas transações com banco.

+++ database/migrations/2021_03_29_152023_create_documento_dinamicos_table.php
@@ -18,11 +18,11 @@ class CreateDocumentoDinamicosTable extends Migration
             $table->unsignedBigInteger('edital_dinamico_tipo_anexo_id');
             $table->string('nome_documento');
             $table->boolean('obrigatorio')->default(false);
-            $table->float('pontuacao_maxima')->nullable();
-            $table->float('pontuacao_por_item')->nullable();
+            $table->double('pontuacao_maxima')->nullable();
+            $table->double('pontuacao_por_item')->nullable();
             $table->integer('quantidade_anexos')->nullable();
-            $table->float('pontuacao_por_ano')->nullable();
-            $table->float('pontuacao_por_mes')->nullable();
+            $table->double('pontuacao_por_ano')->nullable();
+            $table->double('pontuacao_por_mes')->nullable();
             $table->boolean('tipo_experiencia')->nullable();;
             $table->boolean( 'pontuacao_manual')->nullable();;
